rpi/output-service/requirements.txt
nats-py==2.3.1
redis==5.0.1